requirements.txt
backcall==0.1.0
cassandra-driver==3.14.0
decorator==4.3.0
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
kafka==1.3.5
kafka-python==1.4.3
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.5
ptyprocess==0.5.2
Pygments==2.2.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.4
wcwidth==0.1.7
